JS file of the project completed.
This file contains the API key of mapbox. We had entered many things in
it, like the zoom-in and zoom-out feature on the map, we had added this
feature by the help of documentation of mapbox. We had also added a
section by which we can get direction of anywhere, this was the plugin
avaliable on the mapbox website itself.