Readme file updated
By this commit I had updated the README file of this project. This
commit contains many things, first a short summary about the project is
added, then features of this project are added which tells that what we
this app can do, after that I had added a link where you can see the
hosted project, then in last a section is added which tells how to clone
this repository to your local machine.